Change the API for coordination
