+++ app/src/main/java/com/github/geohunt/app/ComposeActivity.kt
@@ -11,9 +11,9 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
-import com.github.geohunt.app.database.Database
-import com.github.geohunt.app.database.DatabaseFactory
-import com.github.geohunt.app.database.firebase.FirebaseDatabase
+import com.github.geohunt.app.model.database.Database
+import com.github.geohunt.app.model.database.DatabaseFactory
+import com.github.geohunt.app.model.database.firebase.FirebaseDatabase
 import com.github.geohunt.app.ui.CreateChallengeView
 import com.github.geohunt.app.ui.theme.GeoHuntTheme
