--- app/src/main/java/com/github/geohunt/app/database/models/PictureImage.kt
@@ -1,13 +1,0 @@
-package com.github.geohunt.app.database.models
-
-import android.graphics.Bitmap
-import java.util.concurrent.CompletableFuture
-
-interface PictureImage {
-    val iid : String
-    val bitmap : Bitmap?
-
-    fun load() : CompletableFuture<Bitmap>
-
-    fun save() : CompletableFuture<Void>
-}
