--- app/src/main/java/com/github/geohunt/app/ui/UIUtility.kt
@@ -1,14 +1,0 @@
-package com.github.geohunt.app.ui
-
-import android.app.Activity
-import android.content.Context
-import android.content.ContextWrapper
-
-fun Context.findActivity() : Activity {
-    var context = this
-    while (context is ContextWrapper) {
-        if (context is Activity) return context
-        context = context.baseContext
-    }
-    throw IllegalStateException("Context.findActivity should be called in the context of an Activity")
-}
